+++ sqleto/CHANGELOG.md
@@ -5,3 +5,9 @@
 ## 0.0.2
 
 - Some few adjusts
+
+## 0.0.3
+
+- Bug fixes on generate scrit to INSERT
+- SQLetoValidator added (more validators soon)
+- Added an example.dart that will be updated when new features coming